pyproject.toml
[project]
name = "pile"
version = "2.0.0"
description = "Local Scrum Master assistant with Jira & Git integration."
license = "MIT"
authors = [{name = "Tan Tran", email = "me@tantran.dev"}]
requires-python = ">=3.11"

dependencies = [
    "agent-framework-ollama",
    "agent-framework-openai>=1.0.0",
    "agent-framework-orchestrations>=1.0.0b260402",
    "chainlit>=2.10.1",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "chromadb>=0.6",
    "pymupdf>=1.25",
    "playwright>=1.40",
    "plotly>=5.0",
]

[project.optional-dependencies]
ui = ["chainlit>=2.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
pile = "pile.ui.cli:main"
pile-web = "pile.ui.chainlit_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pile"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
